refactor: 관심사 따라 web과 core 패키지로 분리
