+++ src/main/java/roomescape/core/domain/ReservationTimeRepository.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.core.domain;
 
 import java.util.List;
 import java.util.Optional;
