+++ src/main/java/roomescape/web/presentation/AdminController.java
@@ -1,4 +1,4 @@
-package roomescape.web;
+package roomescape.web.presentation;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
