+++ src/main/java/roomescape/web/presentation/ReservationController.java
@@ -1,4 +1,4 @@
-package roomescape.web;
+package roomescape.web.presentation;
 
 import java.net.URI;
 import java.util.List;
@@ -10,9 +10,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.service.ReservationService;
-import roomescape.web.dto.ReservationRequest;
-import roomescape.web.dto.ReservationResponse;
+import roomescape.core.service.ReservationService;
+import roomescape.web.presentation.dto.ReservationRequest;
+import roomescape.web.presentation.dto.ReservationResponse;
 
 @RequestMapping("/reservations")
 @RestController
