+++ src/test/java/roomescape/web/integration/IntegrationTest.java
@@ -1,16 +1,18 @@
-package roomescape.integration;
+package roomescape.web.integration;
 
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.ActiveProfiles;
-import roomescape.helper.DatabaseCleaner;
-import roomescape.helper.DatabaseInitializer;
+import roomescape.web.helper.DatabaseCleaner;
+import roomescape.web.helper.DatabaseInitializer;
 
+@EnableAutoConfiguration
 @SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
 @ActiveProfiles("test")
 public abstract class IntegrationTest {
