+++ src/test/java/roomescape/web/integration/ReservationTimeIntegrationTest.java
@@ -1,4 +1,4 @@
-package roomescape.integration;
+package roomescape.web.integration;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.hamcrest.Matchers.is;
